qt countdown test
